Add files via upload
Gaming!!! This update to the files is mainly focused towards enabling gaming support with GPU offloading flags enabled in the amd-nvidia.nix file(make sure to make changes based on your system before using it) as well as enabling Steam and Lutris support in the configuration.nix file. Also included(Although not specific to gaming) are the Kate and Visual Studio Code text editors for development uses.