Update main branch
New configs that include flake and home-manager support, as well as different flake modules for gaming-focused packages like Steam, Lutris etc., Flatpak integration with some flatpaks I personally install on my system to be auto-installed and a multi-drive module that allows other drives in a system to be auto-mounted with read/write/execute permissions(currently setup for one extra drive but can be edited for more). Also converted amd-nvidia.nix into a nix flake for easier modularity. README will be updated in a while.